+++ pkg/conf/computed.go
@@ -10,6 +10,23 @@ import (
 	"github.com/sourcegraph/sourcegraph/schema"
 )
 
+const defaultHTTPStrictTransportSecurity = "max-age=31536000" // 1 year
+
+// HTTPStrictTransportSecurity returns the value of the Strict-Transport-Security HTTP header to set.
+func HTTPStrictTransportSecurity() string {
+	switch v := Get().HttpStrictTransportSecurity.(type) {
+	case string:
+		return v
+	case bool:
+		if !v {
+			return ""
+		}
+		return defaultHTTPStrictTransportSecurity
+	default:
+		return defaultHTTPStrictTransportSecurity
+	}
+}
+
 // SearchTimeoutParameterEnabled returns true if SearchTimeoutParameter experiment is enabled.
 func SearchTimeoutParameterEnabled() bool {
 	p := Get().ExperimentalFeatures.SearchTimeoutParameter